aggregation: split aggregation.py into multiple files
