+++ lir/aggregation/__init__.py
@@ -1,0 +1,20 @@
+from lir.aggregation.base import Aggregation, AggregationData
+from lir.aggregation.case_llr_csv import CaseLLRToCsv
+from lir.aggregation.metrics_csv import WriteMetricsToCsv
+from lir.aggregation.plot import AggregatePlot
+from lir.aggregation.subset import SubsetAggregation
+
+
+__all__ = [
+    # base
+    'AggregationData',
+    'Aggregation',
+    # case_llr_csv
+    'CaseLLRToCsv',
+    # metrics_csv
+    'WriteMetricsToCsv',
+    # plot
+    'AggregatePlot',
+    # subset
+    'SubsetAggregation',
+]
